requirements.txt
# Бот учёта расходов (bot.py) + опционально shift bot (run.py)
# Python 3.10+

# Бот расходов
python-telegram-bot>=21.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0

# Голос (опционально)
SpeechRecognition>=3.10.0
pydub>=0.25.0

# Shift bot (run.py), если нужен
aiogram==3.13.1
aiohttp>=3.9.0
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
